#1985: Make use of MultiButton from @ndla/button
