+++ package.json
@@ -91,25 +91,25 @@
     "@emotion/styled": "^10.0.15",
     "@ndla/accordion": "^0.5.3",
     "@ndla/article-scripts": "^0.5.12",
-    "@ndla/audio-search": "^0.6.26",
-    "@ndla/button": "^0.3.22",
+    "@ndla/audio-search": "^0.6.27",
+    "@ndla/button": "^0.3.23",
     "@ndla/core": "^0.6.21",
     "@ndla/editor": "^0.7.10",
     "@ndla/error-reporter": "^0.4.8",
-    "@ndla/forms": "^0.3.76",
-    "@ndla/howto": "^0.1.84",
+    "@ndla/forms": "^0.3.77",
+    "@ndla/howto": "^0.1.87",
     "@ndla/i18n": "^0.4.2",
     "@ndla/icons": "^0.6.19",
-    "@ndla/image-search": "^0.8.80",
+    "@ndla/image-search": "^0.8.82",
     "@ndla/licenses": "^0.6.22",
-    "@ndla/modal": "^0.4.7",
-    "@ndla/notion": "^0.2.29",
+    "@ndla/modal": "^0.4.8",
+    "@ndla/notion": "^0.2.30",
     "@ndla/pager": "^0.4.25",
     "@ndla/tabs": "^0.11.24",
     "@ndla/tracker": "^0.4.0",
-    "@ndla/ui": "^0.27.18",
+    "@ndla/ui": "^0.27.21",
     "@ndla/util": "^0.3.6",
-    "@ndla/video-search": "^0.8.27",
+    "@ndla/video-search": "^0.8.28",
     "auth0-js": "^9.11.3",
     "babel-plugin-emotion": "^10.0.16",
     "body-parser": "^1.19.0",
