+++ src/components/SlateEditor/EditorFooter.tsx
@@ -18,7 +18,7 @@ import { Article, PossibleStatuses, Values } from './editorTypes';
 import * as draftApi from '../../modules/draft/draftApi';
 import { formatErrorMessage } from '../../util/apiHelpers';
 import { TranslateType } from '../../interfaces';
-import SaveButtonDropDown from '../SaveButtonWithDropDown/SaveButtonWithDropdown';
+import SaveMultiButton from '../SaveMultiButton';
 
 interface Props {
   t: TranslateType;
@@ -81,18 +81,14 @@ const EditorFooter: React.FC<Props> = ({
   }, [values.status]);
 
   const saveButton = (
-    <>
-      <SaveButtonDropDown
-        data-testid="saveLearningResourceButton"
-        isSaving={isSubmitting}
-        defaultText="save"
-        formIsDirty={formIsDirty}
-        large
-        showSaved={savedToServer && !formIsDirty}
-        onClick={onSaveClick}
-        setFieldValue={setFieldValue}
-      />
-    </>
+    <SaveMultiButton
+      large
+      data-testid="saveLearningResourceButton"
+      isSaving={isSubmitting}
+      formIsDirty={formIsDirty}
+      showSaved={savedToServer && !formIsDirty}
+      onClick={onSaveClick}
+    />
   );
 
   const onValidateClick = async () => {
